refactor(styler-plugin): decompose colorUtils.ts into focused modules
Closes #1230

Split colorUtils.ts (646 lines) into 5 focused modules under
colorUtils/ directory:

- types.ts: ColorVariationOptions, ColorVariations interfaces
- colorConversion.ts: HSV/RGB/Hex conversion, adjustBrightness,
  createColorVariations, getContrastRatio
- colorNormalization.ts: normalization algorithms (linear, log,
  quantile, jenks, equal), clamp01, resolveClassCount
- colorCalculation.ts: calculateLinearColor, normalizeStylerConfig,
  generateColorGradient, valueToColor, calculateQuantileColor
- index.ts: barrel re-export preserving public API

typecheck: pass, tests: 77/77 pass
