+++ plugins/styler-plugin/src/index.ts
@@ -41,10 +41,7 @@ export { PLUGIN_MANIFEST as StylerPluginManifest } from './plugin-manifest.js';
 
 // Entity handler
 export { StylerEntityHandler } from './common/handlers/StylerEntityHandler.js';
-export type {
-  ColorVariationOptions,
-  ColorVariations,
-} from './common/utils/colorUtils.js';
+export type { ColorVariationOptions, ColorVariations } from './common/utils/colorUtils/index.js';
 
 // Utilities
 export {
@@ -59,7 +56,7 @@ export {
   rgbToHex,
   rgbToHsv,
   valueToColor,
-} from './common/utils/colorUtils.js';
+} from './common/utils/colorUtils/index.js';
 // Services
 export { StylerDataService } from './services/StylerDataService.js';
